# Conflicts:
#	DirectX11_Starter_2015/Debug/DirectX11_Starter.ilk
#	DirectX11_Starter_2015/Debug/DirectX11_Starter.pdb
#	DirectX11_Starter_2015/Debug/PixelShader.cso
#	DirectX11_Starter_2015/Debug/VertexShader.cso
#	DirectX11_Starter_2015/DirectX11_Starter.sdf
#	DirectX11_Starter_2015/DirectX11_Starter/Debug/Camera.obj
#	DirectX11_Starter_2015/DirectX11_Starter/Debug/Component.obj
#	DirectX11_Starter_2015/DirectX11_Starter/Debug/DirectX1.FEB50FC0.tlog/CL.read.1.tlog
#	DirectX11_Starter_2015/DirectX11_Starter/Debug/DirectX1.FEB50FC0.tlog/CL.write.1.tlog
#	DirectX11_Starter_2015/DirectX11_Starter/Debug/DirectX1.FEB50FC0.tlog/DirectX11_Starter.write.1u.tlog
#	DirectX11_Starter_2015/DirectX11_Starter/Debug/DirectX11_Starter.log
#	DirectX11_Starter_2015/DirectX11_Starter/Debug/DirectXGameCore.obj
#	DirectX11_Starter_2015/DirectX11_Starter/Debug/DrawnMesh.obj
#	DirectX11_Starter_2015/DirectX11_Starter/Debug/Entity.obj
#	DirectX11_Starter_2015/DirectX11_Starter/Debug/EntitySystem.obj
#	DirectX11_Starter_2015/DirectX11_Starter/Debug/Light.obj
#	DirectX11_Starter_2015/DirectX11_Starter/Debug/Material.obj
#	DirectX11_Starter_2015/DirectX11_Starter/Debug/Mesh.obj
#	DirectX11_Starter_2015/DirectX11_Starter/Debug/MyDemoGame.obj
#	DirectX11_Starter_2015/DirectX11_Starter/Debug/Render.obj
#	DirectX11_Starter_2015/DirectX11_Starter/Debug/Resources.obj
#	DirectX11_Starter_2015/DirectX11_Starter/Debug/SimpleShader.obj
#	DirectX11_Starter_2015/DirectX11_Starter/Debug/Transform.obj
#	DirectX11_Starter_2015/DirectX11_Starter/Debug/dxerr.obj
#	DirectX11_Starter_2015/DirectX11_Starter/Debug/vc140.idb
#	DirectX11_Starter_2015/DirectX11_Starter/Debug/vc140.pdb

+++ DirectX11_Starter_2015/DirectX11_Starter/SimpleShader.cpp
@@ -259,7 +259,7 @@ SimpleShaderVariable* ISimpleShader::FindVariable(std::string name, int size)
 
 SimpleShaderVariable* ISimpleShader::FindVariable(int i, int size)
 {
-	if (i < 0 || i >= varTableINT.size())
+	if (i < 0 || i >= (int)varTableINT.size())
 		return 0;
 
 	SimpleShaderVariable* var = &varTableINT[i];
@@ -291,7 +291,7 @@ SimpleConstantBuffer* ISimpleShader::FindConstantBuffer(std::string name)
 
 SimpleConstantBuffer * ISimpleShader::FindConstantBuffer(int i)
 {
-	if (i < 0 || i >= cbTableINT.size())
+	if (i < 0 || i >= (int)cbTableINT.size())
 		return 0;
 	return cbTableINT[i];
 }
@@ -532,7 +532,7 @@ const SimpleSRV* ISimpleShader::GetShaderResourceViewInfo(std::string name)
 
 const SimpleSRV * ISimpleShader::GetShaderResourceView(int index)
 {
-	if (index < 0 || index >= textureTableINT.size()) return nullptr;
+	if (index < 0 || index >= (int)textureTableINT.size()) return nullptr;
 	return textureTableINT[index];
 }
